+++ CHANGELOG.md
@@ -12,6 +12,17 @@ Removed
 
 Fixed
 
+## v3.7.1 - 2016-06-17
+---
+Added
+- validations of input types before cover check
+- cover_types and seal_types to _CONTAINER_TYPES
+Changed
+
+Removed
+- string input types for source, destination wells for Instructions Consolidate, Autopick, Mix
+Fixed
+
 ## v3.7.0 - 2016-06-14
 ---
 Added